Solution to Count Students Scoring Above a Threshold

+++ Count Students Scoring Above a Threshold@@67877d61dc107486ad5c2747/code.c
@@ -11,7 +11,7 @@ int main() {
     for(int i=0;i<n;i++){
         scanf("%d %s %f",&s[i].a,&s[i].b,&s[i].c);
     }
-    int m;
+    float m;
     scanf("%d",&m);
     int count=0;
     for(int i=0;i<n;i++){
